+++ ConnectionHandler/IConnectionHandler.h
@@ -24,14 +24,6 @@ struct Connection {
     std::atomic_int status;
 };
 
-struct threadContext{
-
-    threadContext():  guard(boost::asio::make_work_guard(ioContext)), _thread(boost::bind(&boost::asio::io_service::run, &ioContext)) {}
-
-    boost::asio::io_context ioContext;
-    boost::asio::executor_work_guard<boost::asio::io_context::executor_type> guard;
-    boost::thread _thread;
-};
 
 class IConnectionHandler {
 public:
