+++ InternalEntities/InternalEntities.h
@@ -11,6 +11,8 @@
 #include <boost/archive/binary_iarchive.hpp>
 #include <boost/archive/binary_oarchive.hpp>
 #include <boost/serialization/vector.hpp>
+#include <boost/bind.hpp>
+#include <boost/asio.hpp>
 
 enum class requestType {
     INSERT = 0,
@@ -33,6 +35,8 @@ enum class Status { NOTHING = 0,
                     WORKING,
                     OK,
                     FAILURE };
+
+const std::string delimiter = "\r\n\r\n";
 
 class Request {
 public:
@@ -125,6 +129,13 @@ void Request::load(const std::string &str_data) {
     boost::archive::binary_iarchive ia(iss);
     ia &*(this);
 }
+
+struct threadContext{
+    threadContext():  guard(boost::asio::make_work_guard(ioContext)), _thread(boost::bind(&boost::asio::io_service::run, &ioContext)) {}
+    boost::asio::io_context ioContext;
+    boost::asio::executor_work_guard<boost::asio::io_context::executor_type> guard;
+    boost::thread _thread;
+};
 
 BOOST_CLASS_VERSION(Request, 1)
