+++ ServerDB/main.cpp
@@ -3,7 +3,7 @@
 #include "Initializer.h"
 
 int main() {
-    Initializer dataBase(1);
+    Initializer dataBase(3);
     dataBase.startDB();
 
     std::cout << std::thread::hardware_concurrency() << std::endl;
